requirements.txt
pandas==1.4.2
scikit-learn==1.1.2
imbalanced-learn==0.9.1
hyperopt==0.2.7
mlflow==1.27.0
uvicorn==0.19.0
fastapi==0.85.1
joblib
passlib
PyJWT
python-multipart